+++ AppQuiz.UnitTests/QuizTests/UpdateQuizCommandHandlerTests.cs
@@ -50,7 +50,7 @@ namespace AppQuiz.UnitTests.QuizTests
             var command = new UpdateQuizCommand
             {
                 Title = "anonymousQuizTitle",
-                Priority = Priority.Low,
+                Priority = (int)Priority.Low,
                 ChapterId = Guid.NewGuid(),
                 Questions = new List<Question>() {
                     new Question()
@@ -87,7 +87,7 @@ namespace AppQuiz.UnitTests.QuizTests
             var command = new UpdateQuizCommand
             {
                 Title = "anonymousQuizTitle",
-                Priority = Priority.Low,
+                Priority = (int)Priority.Low,
                 ChapterId = Guid.NewGuid(),
                 Questions = new List<Question>() {
                     new Question()
@@ -125,7 +125,7 @@ namespace AppQuiz.UnitTests.QuizTests
             var command = new UpdateQuizCommand
             {
                 Title = "anonymousQuizTitle",
-                Priority = Priority.Low,
+                Priority = (int)Priority.Low,
                 ChapterId = Guid.NewGuid(),
                 Questions = new List<Question>() {
                     new Question()
@@ -164,7 +164,7 @@ namespace AppQuiz.UnitTests.QuizTests
             var command = new UpdateQuizCommand
             {
                 Title = "anonymousQuizTitle",
-                Priority = Priority.Low,
+                Priority = (int)Priority.Low,
                 Questions = new List<Question>() {
                     new Question()
                     {
